添加deepin stable版

+++ install.sh
@@ -1,3 +1,16 @@
 #!/bin/bash
-sudo cp `dirname $0`/libHTYCMDU.so /usr/lib/dde-dock/plugins/libHTYCMDU.so
-killall dde-dock
+
+codename=$(lsb_release --codename | awk '{print$2}')
+deepin="no"
+if [ $codename=="stable" ]; then
+    deepin="yes"
+elif [ $codename=="unstable" ]; then
+    deepin="yes"
+else
+    echo "未能确定版本"
+fi
+
+if [ "$deepin" = "yes" ]; then
+    sudo cp `dirname $0`/$codename/libHTYCMDU.so /usr/lib/dde-dock/plugins/libHTYCMDU.so
+    killall dde-dock
+fi
